refactor: field types
